requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.4
black==25.1.0
cfgv==3.4.0
click==8.2.1
crispy-bootstrap5==2025.4
distlib==0.3.9
Django==4.2.21
django-bootstrap-v5==1.0.11
django-crispy-forms==2.4
django-excel-response2==3.0.6
django-filter==25.1
django-six==1.0.5
excel-base==1.0.4
filelock==3.18.0
flake8==7.2.0
identify==2.6.10
isort==6.0.1
isoweek==1.3.3
mccabe==0.7.0
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
pre_commit==4.2.0
pycodestyle==2.13.0
pyflakes==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
screen==1.0.1
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
TimeConvert==3.0.13
tomli==2.2.1
typing_extensions==4.13.2
tzlocal==5.3.1
virtualenv==20.31.2
xlwt==1.3.0
